% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affinity_alpha.R
\name{affalpha}
\alias{affalpha}
\title{Affinity by Alpha-Decay Kernel}
\usage{
affalpha(X, nbdk = 5, alpha = 2, ...)
}
\arguments{
\item{X}{an \eqn{(m\times p)} matrix of row-stacked observations of S3 \code{dist} object of \eqn{m} observations.}

\item{nbdk}{neighborhood size to define data-driven bandwidth parameter (default: 5).}

\item{alpha}{decay parameter (default: 2).}

\item{...}{extra parameters including \describe{
\item{zero.diag}{a logical; set the diagonal entries as zeros (default: \code{FALSE}).}
}}
}
\value{
an \eqn{(m\times m)} affinity matrix of \code{"affinity"} class.
}
\description{
\insertCite{moon_visualizing_2019}{dglearn} proposed a data-driven kernel called alpha-decay kernel 
that uses a user-specified decay parameter as well as the nearest-neighbor distance. 
Let \eqn{\sigma_i} be the distance from a point \eqn{x_i} to its \code{nnbd}-th 
nearest neighbor. Then the affinity matrix using the alpha-decay kernel is defined as
\deqn{A_{ij} = \frac{1}{2}
\exp\left(-\left(\frac{\|x_i - x_j\|}{\sigma_i}\right)^\alpha\right) + 
\frac{1}{2}\exp\left(-\left(\frac{\|x_i - x_j\|}{\sigma_j}\right)^\alpha\right).}
}
\examples{
\donttest{
## load 'iris' data
data("iris")

## extract the numerical part
iris_mat = as.matrix(iris[,1:4])

## compute affinity using different neighbor sizes
aff1 = affalpha(iris_mat, nbdk=5)
aff2 = affalpha(iris_mat, nbdk=10)
aff3 = affalpha(iris_mat, nbdk=20)

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,3), pty="s")
image(aff1, xaxt="n", yaxt="n", xlab="", ylab="", main='nbdk=5')
image(aff2, xaxt="n", yaxt="n", xlab="", ylab="", main='nbdk=10')
image(aff3, xaxt="n", yaxt="n", xlab="", ylab="", main='nbdk=20')
par(opar)
}

}
\references{
\insertAllCited{}
}
\concept{affinity}
