% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_dist4randomForest.R
\name{dist4randomForest}
\alias{dist4randomForest}
\title{Extract Distance from Random Forest Objects}
\usage{
dist4randomForest(rfobj, newdata, dtype = c("proximity", "geodesic"))
}
\description{
Extract Distance from Random Forest Objects
}
\concept{interface}
