% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsm_dsmdist2.R
\name{dsmdist2}
\alias{dsmdist2}
\title{Diffusion-based Similarity Metric between Two Stochastic Matrices}
\usage{
dsmdist2(P1, P2)
}
\arguments{
\item{P1}{an \eqn{(m\times m)} row-stochastic matrix.}

\item{P2}{an \eqn{(m\times m)} row-stochastic matrix.}
}
\value{
the DSM value.
}
\description{
Given two \eqn{(m\times m)} row-stochastic matrices with nonnegative entries, 
compute the diffusion-based similarity metric (DSM).
}
\examples{
\donttest{
## load 'iris' and 'USArrests' data
data("iris")
data("USArrests")

## extract the numerical part and select 25 random objects with scaling
dat_i = as.matrix(scale(iris[sample(1:150, 25, replace=F),1:4]))
dat_u = as.matrix(scale(USArrests[sample(1:50, 25, replace=F),]))

## simple affinity and Markov kernels
aff_i = exp(-as.matrix(stats::dist(dat_i))^2)
aff_u = exp(-as.matrix(stats::dist(dat_u))^2)

ker_i = aff_i/base::rowSums(aff_i)
ker_u = aff_u/base::rowSums(aff_u)

## compute the distance
print(paste0("DSM value is ",round(dsmdist2(ker_i, ker_u),4)))
}

}
\concept{dsm}
