% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation_matpower.R
\name{matpower}
\alias{matpower}
\title{Matrix Power}
\usage{
matpower(X, n)
}
\arguments{
\item{X}{an \eqn{(m\times m)} square matrix.}

\item{n}{the number of times a matrix is multiplied. It should be a finite nonnegative integer.}
}
\value{
an \eqn{(m\times m)} square matrix for \eqn{X^n}.
}
\description{
Given a square matrix \eqn{X}, compute its matrix power
\deqn{X^n = X\times \cdots \times X,}
where the multiplication is applied \eqn{n} times. This is an equivalent to 
\code{matrix_power} in Python's numpy linear algebra module.
}
\examples{
## generate a small row-stochastic matrix
X0 = abs(matrix(stats::rnorm(100), ncol=10))
X1 = X0/base::rowSums(X0)

## take a few different powers
X2 = matpower(X1, 2)
X4 = matpower(X1, 4)

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,3), pty="s")
image(X1, xaxt='n', yaxt='n', main="exponent=1")
image(X2, xaxt='n', yaxt='n', main="exponent=2")
image(X4, xaxt='n', yaxt='n', main="exponent=4")
par(opar)

}
\concept{operation}
