% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding_DM.R
\name{embedDM}
\alias{embedDM}
\title{Embedding by Diffusion Maps}
\usage{
embedDM(A, ndim = 2, tsteps = 1)
}
\description{
Embedding by Diffusion Maps
}
\examples{
\donttest{
## load 'iris' data
data("iris")

## extract the numerical part and label
iris_mat = as.matrix(iris[,1:4])
iris_lab = as.factor(iris[,5])

## compute affinity with automatic Gaussian kernel
iris_aff = dglearn::affgaussian(iris_mat)

## run diffusion maps at different step sizes
dm1 = embedDM(iris_aff, tsteps=1)$Y
dm2 = embedDM(iris_aff, tsteps=2)$Y
dm3 = embedDM(iris_aff, tsteps=5)$Y

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,3), pty="s")
plot(dm1, col=iris_lab, cex=0.5, pch=19, asp=1, main="tsteps=1")
plot(dm2, col=iris_lab, cex=0.5, pch=19, asp=1, main="tsteps=2")
plot(dm3, col=iris_lab, cex=0.5, pch=19, asp=1, main="tsteps=5")
par(opar)
}

}
\concept{embedding}
